format: 人工审核公共部分代码提取优化 issue #4655
